Fixed some issues outlined in the excel sheet
