+++ src/components/MxCatalog/Utils/FilterSection.js
