+++ src/components/TrackInfo/TrackHeader.js
@@ -56,7 +56,7 @@ export default function TrackHeader({
 								<a target="_blank" rel="noopener noreferrer">
 									<div className="flex items-end mb-2 font-secondary text-sm">
 										<span className="mr-2">Artwork by-</span>
-										{artworkArtistInfo.userInfo[0] && (
+										{artworkArtistInfo.userInfo[0] && artworkArtistInfo.userInfo[0].avatar && (
 											<Image src={artworkArtistInfo.userInfo[0].avatar} height="20" width="20" className="rounded-full" />
 										)}
 										<span className="ml-1">@{artworkArtistInfo.user[0].username}</span>
@@ -68,7 +68,7 @@ export default function TrackHeader({
 								<a target="_blank" rel="noopener noreferrer">
 									<div className="flex items-end mb-2 font-secondary text-sm">
 										<span className="mr-2">Artwork by-</span>
-										{artworkArtistInfo.userInfo[0] && (
+										{artworkArtistInfo.userInfo[0] && artworkArtistInfo.userInfo[0].avatar && (
 											<Image src={artworkArtistInfo.userInfo[0].avatar} height="20" width="20" className="rounded-full" />
 										)}
 										<span className="ml-1">@{artworkArtistInfo.username}</span>
