+++ styles/NFTCard/Nftcard.module.css
@@ -1,9 +1,9 @@
 .nft-card {
-	@apply flex flex-col overflow-hidden rounded-2xl hover:shadow-2xl duration-300 cursor-pointer;
+	@apply flex flex-col overflow-hidden rounded-2xl hover:shadow-2xl duration-300 cursor-pointer w-[240px];
 }
 
 .nft-card-dark {
-	@apply flex flex-col overflow-hidden rounded-2xl hover:shadow-2xl duration-300 cursor-pointer hover:shadow-neutral-900;
+	@apply flex flex-col overflow-hidden rounded-2xl hover:shadow-2xl duration-300 cursor-pointer hover:shadow-neutral-900 w-[240px];
 }
 
 .nft-card__description {
